requirements.txt
attrs==22.1.0
certifi==2021.5.30
charset-normalizer==2.0.12
colorama==0.4.5
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
importlib-resources==5.4.0
Jinja2==3.0.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-dataclass==8.5.9
matplotlib==3.3.4
missingno==0.5.1
multimethod==1.5
mypy-extensions==0.4.3
networkx==2.5.1
numpy==1.19.5
pandas==1.1.5
pandas-profiling==3.1.0
phik==0.12.0
Pillow==8.4.0
pydantic==1.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.1.1
PyYAML==6.0
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
tangled-up-in-unicode==0.1.0
threadpoolctl==3.1.0
tqdm==4.64.1
types-dataclasses==0.6.3
typing-inspect==0.8.0
typing_extensions==4.1.1
urllib3==1.26.12
visions==0.7.4
wincertstore==0.2
zipp==3.6.0
